refactor: Rename story_machine module to story_generator
Renames the `story_machine` module to `story_generator` to better reflect its purpose.

Updates all imports and references to the module across the codebase, including in the CLI and tests.
